Fixing active track list leak
