+++ src/AmbientSounds.Uwp/Controls/SleepTimerControl.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using AmbientSounds.ViewModels;
 using Microsoft.Extensions.DependencyInjection;
+using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 
 #nullable enable
@@ -12,7 +13,15 @@ namespace AmbientSounds.Controls
         {
             this.InitializeComponent();
             this.DataContext = App.Services.GetRequiredService<SleepTimerViewModel>();
-            this.Unloaded += (_, _) => { ViewModel.Dispose(); };
+            this.Unloaded += OnUnloaded;
+        }
+
+        private void OnUnloaded(object sender, RoutedEventArgs e)
+        {
+            ViewModel.Dispose(); 
+            this.DataContext = null;
+            this.Unloaded -= OnUnloaded;
+            this.Bindings.StopTracking();
         }
 
         public SleepTimerViewModel ViewModel => (SleepTimerViewModel)this.DataContext;
